requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib==1.7.4
argon2-cffi==21.3.0
pydantic
alembic
pydantic-settings
pytz
pydantic[email]
python-multipart
email-validator
jinja2
pre-commit
gunicorn